+++ lib/features/search/presentation/blocs/results_with_matching_address/results_with_matching_address_state.dart
